Update version check for skin second layer
This commit replaces all old version checks with a new attribute called "available_second". This attribute is set to True if the image's version is new and supports a second layer; otherwise, it's False. This change simplifies the code and makes it easier to understand when a skin's second layer is accessible.